+++ sphinx/source/changelog.rst
@@ -23,6 +23,7 @@ Changelog
 * Made the :ref:`audio writer <audioWriter>` a pluggable component
 * Made the :ref:`code map reader <codeMapReader>` a pluggable component
 * Improved handling of TC2048, TC2068 and TS2068 Z80 snapshots
+* Fixed how :ref:`trace.py` builds a PNG image in 128K mode
 * Fixed the bug that prevents :ref:`skool2asm.py`, :ref:`skool2bin.py` and
   :ref:`skool2html.py` from handling an instruction that crosses the 64K
   boundary